+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 
 ### Added
 
+- Characteristic Advancement. (#707)
 - Expanded end of encounter prompt to reset surges. (#931)
 - New compendium content:
   - Level 2 & 3 abilities and features for all classes
@@ -36,6 +37,10 @@
 - Adjusted i18n structure for Power Roll Effects.
 - Removed the limitation on one non-minion creature in Squad combat groups. (#1040)
   - Added controls to the combatants' context menu to toggle whether that monster is the captain or not.
+
+### Removed
+
+- Removed minimum and maximums on characteristic values. (#707)
 
 ### Fixed